extracted find_user to different module
